feat: 컴포넌트 분리, 타입 분리 (#21)
