+++ src/layouts/main/main.styled.ts
@@ -15,7 +15,7 @@ export const Main = styled.main`
 export const MainContainer = styled.div`
 	width: 100%;
 	max-width: 1480px;
-	padding: 0 20px;
-	margin: 0 auto;
+	padding: 0 240px;
+	margin: 20px auto 0 auto;
 	overflow: hidden;
 `;
