+++ src/pages/salary/Salay.styled.ts
@@ -1,58 +1,22 @@
 import styled from "styled-components";
-import { getBorderRadius, getFontSize, getFontWeight } from "@/styles/theme";
-
-export const Container = styled.div`
-	margin: 15px auto;
-	padding: 0 280px;
-	font-family: inherit;
-`;
-
-export const PayButton = styled.button`
-	font-family: inherit;
-	padding: 8px 16px;
-	font-size: ${getFontSize("sm")};
-	color: white;
-	background-color: #029688;
-	border: 1px solid;
-	border-radius: ${getBorderRadius("sm")};
-	cursor: pointer;
-
-	&:hover {
-		background-color: #027056;
-	}
-`;
-
-export const TimeButton = styled.button`
-	font-family: inherit;
-	padding: 8px 16px;
-	font-size: ${getFontSize("sm")};
-	background-color: #f3f4f6;
-	color: black;
-	border: 1px solid;
-	border-radius: ${getBorderRadius("sm")};
-	cursor: pointer;
-
-	&:hover {
-		background-color: #dfe4ea;
-	}
-`;
-export const Title = styled.h1`
-	font-size: 2.8rem;
-	font-weight: ${getFontWeight("bold")};
-	margin-top: 15px;
-`;
+import {
+	getBorderRadius,
+	getColor,
+	getFontSize,
+	getFontWeight,
+} from "@/styles/theme";
 
 export const UserInfo = styled.div`
 	display: flex;
 	justify-content: space-between;
 	width: 100%;
-	padding-top: 10px;
+	padding-top: 15px;
 `;
 
 export const Text = styled.p`
 	margin-top: 5px;
 	font-size: ${getFontSize("md")};
-	color: #637381;
+	color: ${getColor("secondary")};
 `;
 
 export const Section = styled.section`
@@ -62,13 +26,6 @@ export const Section = styled.section`
 export const SectionTitle = styled.h2`
 	font-size: ${getFontSize("xl")};
 	margin-bottom: 15px;
-`;
-
-export const SectionHeader = styled.div`
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 5px;
 `;
 
 export const ActualPayContainer = styled.div`
@@ -91,8 +48,43 @@ export const RightContainer = styled.div`
 export const ActualPayValue = styled.h1`
 	font-size: ${getFontSize("xl")};
 	color: #333;
-	padding-right: 40px;
+	padding-right: 30px;
 	font-weight: ${getFontWeight("medium")};
+`;
+export const PayButton = styled.button`
+	font-family: inherit;
+	padding: 8px 18px;
+	font-size: ${getFontSize("sm")};
+	color: white;
+	background-color: ${getColor("primary")};
+	border: none;
+	border-radius: ${getBorderRadius("sm")};
+	cursor: pointer;
+
+	&:hover {
+		background-color: ${getColor("primary_hover")};
+	}
+`;
+export const SectionHeader = styled.div`
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	margin-bottom: 5px;
+`;
+export const TimeButton = styled.button`
+	font-family: inherit;
+	padding: 8px 16px;
+	font-size: ${getFontSize("sm")};
+	background-color: transparent;
+	color: black;
+	border: 1px solid;
+	border-radius: ${getBorderRadius("sm")};
+	margin-bottom: 8px;
+	cursor: pointer;
+
+	&:hover {
+		background-color: #dfe4ea;
+	}
 `;
 
 export const Tooltip = styled.div`
@@ -103,7 +95,7 @@ export const Tooltip = styled.div`
 	background-color: white;
 	color: black;
 	padding: 10px 15px;
-	border: 1px solid #ccc;
+	border: 1px solid;
 	border-radius: ${getBorderRadius("sm")};
 	font-size: ${getFontSize("xs")};
 	box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
@@ -121,12 +113,11 @@ export const TooltipRow = styled.div`
 	justify-content: space-between;
 	margin-bottom: 5px;
 	font-size: ${getFontSize("xs")};
-	color: #333;
+	color: black;
 
 	&:last-child {
 		margin-bottom: 0;
 	}
-
 	strong {
 		font-weight: ${getFontWeight("bold")};
 	}
@@ -139,15 +130,6 @@ export const TooltipDivider = styled.div`
 	margin: 10px 0;
 `;
 
-export const TimeButtonWrapper = styled.div`
-	position: relative;
-
-	&:hover ${Tooltip} {
-		visibility: visible;
-		opacity: 1;
-	}
-`;
-
 export const List = styled.ul`
 	list-style: none;
 	padding: 0;
@@ -158,7 +140,7 @@ export const ListItem = styled.li`
 	display: flex;
 	justify-content: space-between;
 	align-items: center;
-	margin-bottom: 10px;
+	margin-bottom: 12px;
 	font-size: ${getFontSize("md")};
 	border-bottom: 2px solid transparent;
 
@@ -167,14 +149,14 @@ export const ListItem = styled.li`
 	}
 
 	&:hover {
-		border-bottom: 2px solid #dfe4ea;
+		border-bottom: 2px solid ${getColor("secondaryLight")};
 	}
 `;
 
 export const Label = styled.p`
 	margin: 0;
 	font-size: ${getFontSize("md")};
-	color: #637381;
+	color: ${getColor("secondary")};
 	padding-bottom: 5px;
 `;
