+++ src/pages/salary/components/Header.styled.ts
