+++ src/pages/salary/types/Graph.ts
@@ -1,0 +1,4 @@
+export interface GraphProps {
+	salaryPercent: number;
+	deductionPercent: number;
+}
