+++ src/pages/salary/types/SalaryDetailSection.ts
@@ -1,0 +1,7 @@
+interface DeductionItem {
+	label: string;
+	value: string;
+}
+export interface SalaryDetailsSectionProps {
+	deductionItems: DeductionItem[];
+}
