+++ src/pages/salary/types/SalarySection.ts
@@ -1,0 +1,4 @@
+export interface SalarySectionProps {
+	isModalOpen: any;
+	setIsModalOpen: any;
+}
