+++ src/router/index.tsx
@@ -4,8 +4,8 @@ import HomePage from "@/pages/home";
 import Schedule from "@/pages/schedule";
 import LoginPage from "@/pages/login";
 import SignupPage from "@/pages/signup";
-import PayPage from "@/pages/pay";
-import PayCorrectionPage from "@/pages/pay-correction";
+import SalaryPage from "@/pages/salary";
+import SalaryCorrectionPage from "@/pages/salary-correction";
 import MyPage from "@/pages/mypage";
 import NotFoundPage from "@/pages/notFound";
 
@@ -15,8 +15,8 @@ const Router = createBrowserRouter([
 		children: [
 			{ path: "/", element: <HomePage /> },
 			{ path: "/schedule", element: <Schedule /> },
-			{ path: "/pay", element: <PayPage /> },
-			{ path: "/pay-correction", element: <PayCorrectionPage /> },
+			{ path: "/salary", element: <SalaryPage /> },
+			{ path: "/salary-correction", element: <SalaryCorrectionPage /> },
 			{ path: "/mypage", element: <MyPage /> },
 		],
 	},
